agri-monitoring-backend-main/requirements.txt
# Web & API
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.11
dj-database-url==2.1.0

# Environment
python-dotenv==1.0.1
requests==2.32.3

# Data / ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2

# Development tools
black==24.8.0
pylint-django==2.5.5